Implementation/eventservice/requirements.txt
sqlalchemy
Flask
psycopg2
uwsgi
pika
python-dateutil